memd: Revive mem_read()
If there is wrapper script, it is very enough for memd to have simple feature
only such as read/write.
Therefore, in the future, using wrapper script will be mandatory, and then,
DUMP feature will be removed.

Thus, wrapper script should use another function instead of mem_dump().
This patch adds it.

But, if we remove DUMP feature from memd, it breaks backward compatibility
and using wrapper script is mandatory.

Signed-off-by: Yuya Hamamachi <yuya.hamamachi.sx@renesas.com>
Acked-by: Kuninori Morimoto <kuninori.morimoto.gx@renesas.com>
Acked-by: Yoshihiro Shimoda <yoshihiro.shimoda.uh@renesas.com>